+++ NEWS.md
@@ -4,6 +4,10 @@
 
 ## Unreleased (2026-07-12)
 
+#### Breaking changes
+
+- remove the cash and universal harvest resolver wrappers
+
 #### New features
 
 - add hook-based wrapper customizations